ApiViewDumps: return only raw download URL (#54)
This replaces the current "link" response, which returns an HTML element, with a new one that returns only the URL of the dump.

These changes are made with the use-case of a client that manages these dumps through the API in mind. For these clients, currently the only way to retrieve the URL to the dump is to include an HTML parser and try to retrieve it from the href attribute of the a element inside the link response. Implementing these changes would make life easier for them